separated the functionality into a hook form
