+++ capacitor.config.ts
@@ -3,7 +3,7 @@ import type { CapacitorConfig } from '@capacitor/cli';
 const config: CapacitorConfig = {
   appId: 'com.example.app',
   appName: 'react',
-  webDir: 'build'
+  webDir: 'build',
 };
 
 export default config;
